change data-type long to int.

+++ Practice/include/HomeAutomation/LightScheduler.h
@@ -14,7 +14,7 @@ void LightScheduler_Destroy(void);
 
 void LightScheduler_wakeup(void);
 
-void LightScheduler_scheduleTurnOn(int id, DAY day, long minuite_of_day);
-void LightScheduler_scheduleTurnOff(int id, DAY day, long minuite_of_day);
+void LightScheduler_scheduleTurnOn(int id, DAY day, int minuite_of_day);
+void LightScheduler_scheduleTurnOff(int id, DAY day, int minuite_of_day);
 
 #endif  /* D_LightScheduler_H */
